# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Lawcaspro Legal Operations System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List cases
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: integer
        - name: developerId
          in: query
          schema:
            type: integer
        - name: assignedLawyerId
          in: query
          schema:
            type: integer
        - name: purchaseMode
          in: query
          schema:
            type: string
        - name: titleType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseListResponse"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
  /cases/{caseId}/workflow:
    get:
      operationId: getCaseWorkflow
      tags: [cases]
      summary: Get case workflow steps
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowStep"
  /cases/{caseId}/workflow/{stepId}:
    patch:
      operationId: updateWorkflowStep
      tags: [cases]
      summary: Update a workflow step
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStep"
  /cases/{caseId}/notes:
    get:
      operationId: getCaseNotes
      tags: [cases]
      summary: Get case notes
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseNote"
    post:
      operationId: createCaseNote
      tags: [cases]
      summary: Add a note to a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"
...
  /cases/recent:
    get:
      operationId: getRecentCases
      tags: [cases]
      summary: Get recently updated cases
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseSummary"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CasePurchaser:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        icNo:
          type: ["string", "null"]
        role:
          type: string
          description: "main | joint"
        orderNo:
          type: integer
      required:
        - id
        - clientId
        - clientName
        - icNo
        - role
        - orderNo
    CaseAssignment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        roleInCase:
          type: string
          description: "lawyer | clerk"
        assignedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - roleInCase
        - assignedAt
    CaseSummary:
      type: object
      properties:
        id:
          type: integer
        referenceNo:
          type: string
        projectName:
          type: string
        developerName:
          type: string
        purchaseMode:
          type: string
        titleType:
          type: string
        status:
          type: string
        assignedLawyerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referenceNo
        - projectName
        - developerName
        - purchaseMode
        - titleType
        - status
        - assignedLawyerName
        - createdAt
    CaseDetail:
      type: object
      properties:
        id:
          type: integer
        firmId:
          type: integer
        referenceNo:
          type: string
        projectId:
          type: integer
        projectName:
          type: string
        developerId:
          type: integer
        developerName:
          type: string
        purchaseMode:
          type: string
          description: "cash | loan"
        titleType:
          type: string
          description: "master | individual | strata"
        spaPrice:
          type: ["number", "null"]
        status:
          type: string
        purchasers:
          type: array
          items:
            $ref: "#/components/schemas/CasePurchaser"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/CaseAssignment"
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firmId
        - referenceNo
        - projectId
        - projectName
        - developerId
        - developerName
        - purchaseMode
        - titleType
        - spaPrice
        - status
        - purchasers
        - assignments
        - createdBy
        - createdAt
    CaseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CaseSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateCaseBody:
      type: object
      properties:
        projectId:
          type: integer
        developerId:
          type: integer
        purchaseMode:
          type: string
        titleType:
          type: string
        spaPrice:
          type: number
        assignedLawyerId:
          type: integer
        assignedClerkId:
          type: integer
        purchaserIds:
          type: array
          items:
            type: integer
      required:
        - projectId
        - developerId
        - purchaseMode
        - titleType
        - assignedLawyerId
        - purchaserIds
    UpdateCaseBody:
      type: object
      properties:
        status:
          type: string
        purchaseMode:
          type: string
        titleType:
          type: string
        spaPrice:
          type: number
        assignedLawyerId:
          type: integer
        assignedClerkId:
          type: integer

    # ── Workflow ─────────────────────────────────────────────────────────────
    WorkflowStep:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        stepKey:
          type: string
        stepName:
          type: string
        stepOrder:
          type: integer
        status:
          type: string
          description: "pending | in_progress | completed | skipped"
        pathType:
          type: string
          description: "common | loan | mot | noa_pa"
        completedBy:
          type: ["integer", "null"]
        completedByName:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
      required:
        - id
        - caseId
        - stepKey
        - stepName
        - stepOrder
        - status
        - pathType
        - completedBy
        - completedByName
        - completedAt
        - notes
    UpdateWorkflowStepBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
    CaseNote:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - authorId
        - authorName
        - content
        - createdAt
    CreateCaseNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...