# lib/api-zod/src/generated/types/caseDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lawcaspro Legal Operations System API
 * OpenAPI spec version: 1.0.0
 */
import type { CaseAssignment } from "./caseAssignment";
import type { CasePurchaser } from "./casePurchaser";

export interface CaseDetail {
  id: number;
  firmId: number;
  referenceNo: string;
  projectId: number;
  projectName: string;
  developerId: number;
  developerName: string;
  /** cash | loan */
  purchaseMode: string;
  /** master | individual | strata */
  titleType: string;
  /** @nullable */
  spaPrice: number | null;
  status: string;
  purchasers: CasePurchaser[];
  assignments: CaseAssignment[];
  /** @nullable */
  createdBy: number | null;
  createdAt: Date;
}