# lib/api-zod/src/generated/types/caseSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lawcaspro Legal Operations System API
 * OpenAPI spec version: 1.0.0
 */

export interface CaseSummary {
  id: number;
  referenceNo: string;
  projectName: string;
  developerName: string;
  purchaseMode: string;
  titleType: string;
  status: string;
  /** @nullable */
  assignedLawyerName: string | null;
  createdAt: Date;
}