# lib/api-zod/src/generated/types/createCaseNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lawcaspro Legal Operations System API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateCaseNoteBody {
  content: string;
}